requirements.txt
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.6.1
scikit-surprise==1.1.4
pyarrow
joblib 